Using random numbers as weitghs

+++ fitness/number_match.py
@@ -17,12 +17,12 @@ class number_match(base_ff):
         expression = ind.phenotype
         expression_parts = expression.split(" ")
 
-        result = int(expression_parts[0])
+        result = float(expression_parts[0])
 
         for i in range(1, len(expression_parts), 2):
             if expression_parts[i] == "+":
-                result += int(expression_parts[i + 1])
+                result += float(expression_parts[i + 1])
             else:
-                result -= int(expression_parts[i + 1])
+                result -= float(expression_parts[i + 1])
 
         return abs(result - self.target)
